NewEffect: use int for Direction field
(cherry picked from commit a631ab2deb9cf59c8295a75a8286e799d0186cde)

+++ src/npc.cpp
@@ -3290,7 +3290,7 @@ void SpecialNPC(int A)
                         {
                             for(int Ci = 1; Ci <= 10; Ci++)
                             {
-                                NewEffect(EFFID_PLR_FIREBALL_TRAIL, NPC[A].Location, static_cast<float>(NPC[A].Special));
+                                NewEffect(EFFID_PLR_FIREBALL_TRAIL, NPC[A].Location, NPC[A].Special);
                                 Effect[numEffects].Location.SpeedX = dRand() * 3 - 1.5 + NPC[A].Location.SpeedX * 0.1;
                                 Effect[numEffects].Location.SpeedY = dRand() * 3 - 1.5 - NPC[A].Location.SpeedY * 0.1;
